Explicit unimodal and bimodal modulations
We no longer pass background settings. Instead, a boolean controls the state of unimodal or bimodal modulations. The background settings are dynamically calculated whenever this state is changed. This has the effect of fixing a bug in amplitude modulation of bimodal flicker around a non half-on background. Also added method for setting the blink duration.